# Remove branches from isValidBase16