@trask This changed the flamegraphs quite a bit to not have a dominant base16 validation in them. 0 is still better than non-0 but maybe this changes the importance of the other PR (though I left a comment on that which could allow us to eat the cake).

After
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  3208.277 ± 48.574  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60  1205.990 ± 18.235  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   592.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60  1206.781 ± 33.014  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   592.359 ± 13.208    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.001 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   561.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   232.000               ms
```

Before
```
Benchmark                                                          Mode  Cnt     Score    Error   Units
FillSpanBenchmark.setFourAttributes                               thrpt   60  2029.646 ± 25.532  ops/ms
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate                thrpt   60   763.105 ±  9.714  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.alloc.rate.norm           thrpt   60   592.000 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space       thrpt   60   762.296 ± 25.178  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Eden_Space.norm  thrpt   60   591.415 ± 18.165    B/op
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen          thrpt   60     0.002 ±  0.001  MB/sec
FillSpanBenchmark.setFourAttributes:·gc.churn.G1_Old_Gen.norm     thrpt   60     0.002 ±  0.001    B/op
FillSpanBenchmark.setFourAttributes:·gc.count                     thrpt   60   427.000           counts
FillSpanBenchmark.setFourAttributes:·gc.time                      thrpt   60   164.000               ms
```